requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.0
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.0.5
django-addresses==0.1
django-bootstrap-form==3.4
django-crispy-forms==1.9.0
django-extensions==2.2.8
django-filter==2.2.0
django-invoice==0.3.2
django-jsonfield==1.4.0
django-model-utils==4.0.0
django-widget-tweaks==1.4.8
html5lib==1.0.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
ordereddict==1.1
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pycparser==2.20
pylint==2.4.4
PyPDF2==1.26.0
Pyphen==0.9.5
pytz==2019.3
reportlab==3.5.42
signals==0.0.2
six==1.14.0
sqlparse==0.3.1
tinycss2==1.0.2
typed-ast==1.4.1
WeasyPrint==51
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.4
